tests: Make the test_closure test less CPU intensive
As far as I can see the test already had a tendency to timeout and [1]
pushed it over the threshold of 1 second on Travis (Python 2 builds).
I don't see a point of flooding the socket like this so adding an
artifical delay seems like an easy "fix" here.

From what I understand the logic of the test (introduced in [2]) remains
unchanged - a closer greenthread closes the client socket, client
greenthread finishes because of EBADF and server greenthread finishes
because of broken pipe.

[1] bc4d1b5d362e5baaeded35b1e339b9db08172dd2
[2] 1a8a9fc0512299c8fe944784525a021dbd2fe2a0