green selectors: Remove non-green selectors
The code pretending that all the selector classes exist and
"implementing" them using SelectSelector was added in [1] because there
was no facility to easily remove things from green version of modules
(as far as I understand) - the facility has been added in [2] thus
allowing us to make this change.

As mentioned in the code I believe this approach is likely to be less
confusing than the previous one (SelectSelector doesn't expose fileno()
method that the other selectors have so it's not fully API-compatible
with them).

[1] 0d509ef7d2eea3ed4f8ade35d5d09918c811b56c
[2] f63165c0e3c85699ebdb454878d1eaea13e90553